+++ jam/hash.c
@@ -252,6 +252,6 @@ hashstat( struct hash *hp )
 		hp->items.nel,
 		hp->tab.nel,
 		hp->items.nel * hp->items.size / 1024,
-		hp->tab.nel * sizeof( ITEM ** ) / 1024,
+		(int)(hp->tab.nel * sizeof( ITEM ** ) / 1024),
 		(float)count / (float)sets );
 }
